исправила ошибки в A2 из 7 темы
