+++ src/Lizun7/A2/Run.java
@@ -1,0 +1,12 @@
+package Lizun7.A2;
+
+public class Run {
+    public static void main(String[] args) {
+        Parrot parrot = new Parrot() {
+            void say() {
+                System.out.println("А этот попугай умеет говорить!");
+            }
+        };
+        parrot.say();
+    }
+}
